requirements.txt
python-docx==1.1.2
markdown-it-py==3.0.0
PyYAML==6.0.2
requests==2.32.3
latex2mathml==3.77.0
lxml==5.3.0
pytest==8.3.4
